requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
requests
scikit-learn
